Fix dubbo.tag override by consumer.tag in NacosRegistry
